Address navneet1v review: float[]/SIMD/Lombok/javadocs
Refactors the MUVERA encoder and processors per review feedback on PR #3164:

- Convert encoder hot path from double[]/double[][] to float[][] throughout.
  Document/query inputs, cluster centers, projection rows, and SimHash planes
  are now all float-precision.
- Use Lucene's VectorUtil.dotProduct(float[], float[]) for both the SimHash
  partition step and the random projection step so the JVM can emit SIMD/FMA
  instructions on the dot product hot loop.
- Replace List<List<Integer>> cluster index buffers with flat int[][]
  pre-allocated to the worst-case size, eliminating per-loop boxing.
- Lay out each SimHash hyperplane and each projection row as its own
  float[dim] so VectorUtil.dotProduct is called on contiguous arrays without
  slicing or copying inside the loop.
- Switch MuveraSearchRequestProcessor to set the FDE context attribute as a
  raw float[] (no List<Float> boxing). XContentBuilder serializes float[]
  natively on the template-resolution path.
- Annotate processors with @Log4j2 and use @NoArgsConstructor for the utility
  class, matching k-NN conventions.
- Drop the cluster-level Factory log line; the computed FDE dimension is
  surfaced via the auto-generated processor description (visible through
  GET _ingest/pipeline / GET _search/pipeline) which is user-visible.
- Move config string keys into private static final constants.
- Add javadocs to public classes, factories, and overrides.

Tests updated to use float[][] inputs and to assert that the search
processor's context attribute is a float[]. The ingest processor still
materializes a List<Float> in _source because IngestDocument.deepCopy does
not support float[].

+++ src/test/java/org/opensearch/knn/processor/muvera/MuveraEncoderTests.java
@@ -23,7 +23,7 @@ public class MuveraEncoderTests extends KNNTestCase {
         MuveraEncoder enc1 = new MuveraEncoder(4, 1, 2, 2, 42L);
         MuveraEncoder enc2 = new MuveraEncoder(4, 1, 2, 2, 42L);
 
-        double[][] vectors = { { 1.0, 0.0, 0.0, 0.0 }, { 0.0, 1.0, 0.0, 0.0 } };
+        float[][] vectors = { { 1.0f, 0.0f, 0.0f, 0.0f }, { 0.0f, 1.0f, 0.0f, 0.0f } };
 
         float[] fde1 = enc1.processDocument(vectors);
         float[] fde2 = enc2.processDocument(vectors);
@@ -38,7 +38,7 @@ public class MuveraEncoderTests extends KNNTestCase {
         MuveraEncoder enc1 = new MuveraEncoder(4, 1, 2, 2, 42L);
         MuveraEncoder enc2 = new MuveraEncoder(4, 1, 2, 2, 99L);
 
-        double[][] vectors = { { 1.0, 0.0, 0.0, 0.0 }, { 0.0, 1.0, 0.0, 0.0 } };
+        float[][] vectors = { { 1.0f, 0.0f, 0.0f, 0.0f }, { 0.0f, 1.0f, 0.0f, 0.0f } };
 
         float[] fde1 = enc1.processDocument(vectors);
         float[] fde2 = enc2.processDocument(vectors);
@@ -55,7 +55,7 @@ public class MuveraEncoderTests extends KNNTestCase {
 
     public void testDocumentVsQueryModeDiffers() {
         MuveraEncoder encoder = new MuveraEncoder(4, 1, 2, 2, 42L);
-        double[][] vectors = { { 1.0, 0.0, 0.0, 0.0 }, { 0.0, 1.0, 0.0, 0.0 } };
+        float[][] vectors = { { 1.0f, 0.0f, 0.0f, 0.0f }, { 0.0f, 1.0f, 0.0f, 0.0f } };
 
         float[] docFde = encoder.processDocument(vectors);
         float[] queryFde = encoder.processQuery(vectors);
@@ -76,7 +76,7 @@ public class MuveraEncoderTests extends KNNTestCase {
 
     public void testOutputDimensionMatchesExpected() {
         MuveraEncoder encoder = new MuveraEncoder(4, 1, 2, 2, 42L);
-        double[][] vectors = { { 1.0, 2.0, 3.0, 4.0 } };
+        float[][] vectors = { { 1.0f, 2.0f, 3.0f, 4.0f } };
 
         float[] docFde = encoder.processDocument(vectors);
         assertEquals(encoder.getEmbeddingSize(), docFde.length);
@@ -87,7 +87,7 @@ public class MuveraEncoderTests extends KNNTestCase {
 
     public void testSingleVector() {
         MuveraEncoder encoder = new MuveraEncoder(4, 1, 2, 2, 42L);
-        double[][] vectors = { { 1.0, 0.0, 0.0, 0.0 } };
+        float[][] vectors = { { 1.0f, 0.0f, 0.0f, 0.0f } };
 
         float[] fde = encoder.processDocument(vectors);
         assertEquals(8, fde.length);
@@ -127,7 +127,7 @@ public class MuveraEncoderTests extends KNNTestCase {
         MuveraEncoder encoder = new MuveraEncoder(4, 0, 2, 3, 42L);
         assertEquals(3 * 1 * 2, encoder.getEmbeddingSize()); // 6
 
-        double[][] vectors = { { 1.0, 0.0, 0.0, 0.0 }, { 0.0, 1.0, 0.0, 0.0 } };
+        float[][] vectors = { { 1.0f, 0.0f, 0.0f, 0.0f }, { 0.0f, 1.0f, 0.0f, 0.0f } };
         float[] fde = encoder.processDocument(vectors);
         assertEquals(6, fde.length);
     }
